+++ decorators/capitalize_decorator.rb
@@ -1,0 +1,7 @@
+require_relative 'decorators/decorator'
+
+class CapitalizeDecorator < Decorator
+  def correct_name
+    super.capitalize
+  end
+end
