+++ decorators/trimmer_decorator.rb
@@ -1,0 +1,7 @@
+require_relative 'decorators/decorator'
+
+class TrimmerDecorator < Decorator
+  def correct_name
+    super.strip
+  end
+end
