+++ person.rb
@@ -26,26 +26,3 @@ class Person < Nameable
 
   private :of_age?
 end
-
-class Decorator < Nameable
-  def initialize(nameable)
-    super()
-    @nameable = nameable
-  end
-
-  def correct_name
-    @nameable.correct_name
-  end
-end
-
-class CapitalizeDecorator < Decorator
-  def correct_name
-    super.capitalize
-  end
-end
-
-class TrimmerDecorator < Decorator
-  def correct_name
-    super.strip
-  end
-end
